Hand picked some changes from NAudio 1.6 that fix an exception when a USB device is removed while playback is in progress.
On this page http://mark-dot-net.blogspot.co.uk/2012/10/naudio-16-release-notes-10th.html in the "What's new in NAudio 1.6" section, I found the following:
"PlaybackStopped and RecordingStopped events now include an exception. This is very useful for cases when USB audio devices are removed during playback or record. Now there is no unhandled exception and you can detect this has happened by looking at the EventArgs. (n.b. I’m not sure if adding a property to an EventArgs is a breaking change – recompile your code against NAudio 1.6 to be safe)."